+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.12.7+1
- * Change the signature for examples in Intl.plural/gender/select to match
-   message, allowing dynamic values.
+ * Change the signature for args and examples in Intl.plural/gender/select to
+   match Intl.message, allowing dynamic values.
 
 ## 0.12.7
  * Update SDK dependency to 1.12.0, to reflect use of null-aware operators.